Refactoring weather forecast dependencies file structure
