+++ life-manager-web/src/components/weather_forecast/utils/WeatherIcons.style.scss
