+++ life-manager-web/src/components/weather_forecast/utils/weatherConditions.tsx
